feat: add use integers
